Fix #67: The add-on's shading and overlay settings were not applied correctly.
The reason was that prior to the refactoring done for Alice/LG 2.0 there 
was a separate Blender window with a separate SpaceView3D for the 
lightfield window. This was holding all settings. Since this was removed 
during refactoring there was no space which was holding the add-on 
settings. The solution was to create internal dictionaries of the 
settings, which are save the viewport settings before drawing and 
restore them afterwards.