Updated/Fixed: Camera settings options / camera markers
This commit resolves issues #12, #29, #38, and #99.

The addon now respects camera markers set by the user, which allows to change the active camera during rendering of quilt animations.

In addition, the camera selector for the Looking Glass now received a linking option. If this is activated, the selector will automatically select the active camera as the Looking Glass camera. Additionally, the focal plane received a linking option as well, which enables user's to link the focus distance (depth-of-field) of the Blender camera and the focal plane of the Looking Glass settings.